Fix #142: separate out modifiers that don't use integrators
