+++ Source/Waterfall/Waterfall.csproj
@@ -114,6 +114,7 @@
     <Compile Include="EffectControllers\RemapController.cs" />
     <Compile Include="EffectControllers\ThrottleController.cs" />
     <Compile Include="EffectControllers\ThrustController.cs" />
+    <Compile Include="EffectModifiers\DirectModifier.cs" />
     <Compile Include="EffectModifiers\EffectLightColorModifier.cs" />
     <Compile Include="EffectModifiers\EffectLightFloatModifier.cs" />
     <Compile Include="EffectModifiers\EffectColorFromLightModifier.cs" />
